% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv.costs.R
\name{calculate.alpha.cost}
\alias{calculate.alpha.cost}
\title{Intra-Variable Penalised Cost Calculation}
\usage{
calculate.alpha.cost(n, n.nonzero, x, x2, p, alpha, length.checklist,
  cost.func = norm.meanvar)
}
\arguments{
\item{n}{A matrix containing the lengths between the changepoints in current changepoint vector being considered and the changepoints in the check-list of previous changepoint vectors.}

\item{n.nonzero}{A matrix containing the logical values indicating which of the corresponding values in \code{n} are non-zero.}

\item{x}{A \code{p} x \code{length.checklist} matrix containing the cumulative sums of the observations between the current changepoint vector and each corresponding changepoint vector on the check-list.}

\item{x2}{The same as \code{x}, except containing the cumulative sums of the squared observations.}

\item{p}{The number of variables in the series.}

\item{alpha}{The variable-specific penalty, used to penalise the addition of a given changepoint into a given variable. A non-negative numeric value.}

\item{length.checklist}{The number of possible previous changepoint vectors on the check-list being considered.}

\item{cost.func}{The function being used to calculate the multivariate cost. Required to be a function (not a string name).}
}
\value{
A vector of length \code{length.checklist} containing the intra-variable penalised costs for each changepoint vector on the check-list.
}
\description{
Calculates the intra-variable penalised cost (i.e. using only the \code{alpha} penalty) for all variables for all possible previous changepoint vectors.
}
\details{
An internal function, not designed for use by the end-user.
}
