% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptmv.class.R
\docType{class}
\name{cptmv-class}
\alias{cptmv-class}
\alias{cptmv}
\title{cptmv: An S4 class for multivariate changepoints}
\description{
An S4 class used to return information on detected changepoints in multivariate time series.
}
\details{
This S4 class is used to return the results of subset-multivariate changepoint detection methods. The information included relates the detected changepoints, including their locations, the variables which are affected, the likelihood value of the final segmentation. Other information such as the cost function and penalty values used and the running time of the detection procedure are also included.
}
\section{Slots}{

\describe{
\item{\code{data.set}}{A ts object. The multivariate data set being analysed for changepoints.}

\item{\code{cost.func}}{A character object. The name of the function used to calculate the (unpenalised) cost.}

\item{\code{cpt.type}}{A character object. The type of change(s) which are being detected.}

\item{\code{alpha}}{A numeric object. The value of the alpha penalty used.}

\item{\code{beta}}{A numeric object. The value of the beta penalty used.}

\item{\code{num.cpt.vecs}}{A numeric object. The number of changepoint vectors considered by the detection procedure.}

\item{\code{cpt.vecs}}{A numeric matrix. The set of estimated changepoint vectors.}

\item{\code{like}}{A numeric object. The likelihood value of the estimated segmentation.}

\item{\code{cpts}}{A numeric object. The set of estimated changepoint locations.}

\item{\code{subsets}}{A boolean matrix. The sets of affected variables corresponding to each detected changepoint.}

\item{\code{runtime}}{A numeric object. The running time of the detection procedure, in seconds.}
}}

\author{
Benjamin Pickering
}
