% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.functions.R
\name{create.perms}
\alias{create.perms}
\title{Permutation Creation}
\usage{
create.perms(n = 20, p = 3, source.vector = 1:n)
}
\arguments{
\item{n}{The size of the source vector used to generate the permutations.}

\item{p}{The length of the permutations to be produced.}

\item{source.vector}{The vectors of numbers used to generate the permutations.}
}
\value{
perms A matrix containing all of the permutations generated.
}
\description{
Creates all permutations of length \code{p} using the \code{n}-length vector of numbers given by \code{source.vector}.
}
\examples{
create.perms(n=20, p=3) # produces a 20^3 matrix of permutations.
}
