% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mop_distance.R
\name{mop_distance}
\alias{mop_distance}
\title{mop_distance: Estimate mop distance between two matrices}
\usage{
mop_distance(M, G)
}
\arguments{
\item{M}{A numeric matrix of dimension n x m}

\item{G}{A numeric matrix of dimension n x m}
}
\value{
A numeric value that represents the Euclidean distance between the
two matrices
}
\description{
mop_distances calculates the Euclidean distance
between two matrices.
}
\examples{

set.seed(111)
n=1000
M <- matrix(rnorm(4*n), 2*n, 2)
G <- matrix(rnorm(4*n), 2*n, 2)
mdist <- smop::mop_distance(M,G)
}
