% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv.costs.R
\name{norm.meanvar.seglen}
\alias{norm.meanvar.seglen}
\title{Multivariate Cost Calculation: Normally distributed observations, detecting changes in both mean and variance, with additional segment length penalty in likelihood.}
\usage{
norm.meanvar.seglen(n, x, x2, p, length.checklist)
}
\arguments{
\item{n}{A matrix containing the lengths between the changepoints in current changepoint vector being considered and the changepoints in the check-list of previous changepoint vectors.}

\item{x}{A \code{p} x \code{length.checklist} matrix containing the cumulative sums of the observations between the current changepoint vector and each corresponding changepoint vector on the check-list.}

\item{x2}{The same as \code{x}, except containing the cumulative sums of the squared observations.}

\item{p}{The number of variables in the series.}

\item{length.checklist}{The number of possible previous changepoint vectors on the check-list being considered.}
}
\value{
A vector of length \code{length.checklist} containing the multivariate costs for each changepoint vector on the check-list being considered.
}
\description{
Calculates the cost between a sequence of multivariate observations using the provided cumulative sums of data, with the additional penalty of the segment length (\code{log(n)}). The observations are assumed to be Normally distributed. The mean and variance parameters are set to their maximum likelihood estimates. Used for detecting changes in both mean and variance.
}
\details{
An internal function, not designed for use by the end-user.
}
