% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelt.R, R/pelt.bck.R
\name{pelt.exp.cost.seglen}
\alias{pelt.exp.cost.seglen}
\title{Title}
\usage{
pelt.exp.cost.seglen(tau, R, sumx)

pelt.exp.cost.seglen(tau, R, sumx)
}
\arguments{
\item{sumx}{}

\item{sumx}{}
}
\value{
A vector containing the cost of the different possible segments defined by \code{tau} and \code{R}.

A vector containing the cost of the different possible segments defined by \code{tau} and \code{R}.
}
\description{
Title

Title
}
