% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelt.R, R/pelt.bck.R
\name{pelt.gamma.cost}
\alias{pelt.gamma.cost}
\title{Calculate Gamma likelihood of data segment, assuming fixed shape parameter and variable scale parameter.}
\usage{
pelt.gamma.cost(tau, R, sumx, shape = 1)

pelt.gamma.cost(tau, R, sumx, shape = 1)
}
\arguments{
\item{tau}{A vector of locations indicating the possible beginnings of a segment.}

\item{R}{A single location representing the end of the possible segment}

\item{sumx}{The summary statistics for the entire time series being analysed.}

\item{shape}{The value of the shape parameter for the Gamma distribution.}

\item{tau}{A vector of locations indicating the possible beginnings of a segment.}

\item{R}{A single location representing the end of the possible segment}

\item{sumx}{The summary statistics for the entire time series being analysed.}

\item{shape}{The value of the shape parameter for the Gamma distribution.}
}
\value{
A vector containing the cost of the different possible segments defined by \code{tau} and \code{R}.

A vector containing the cost of the different possible segments defined by \code{tau} and \code{R}.
}
\description{
Calculates the Gamma likelihood of multiple segments of data, as defined by the vector of possible start-points (\code{tau}) and a single common end-point (\code{R}). This calculation assumes the shape is fixed as some specified value and sets the scale equal to its maximum likelihood estimate.

Calculates the Gamma likelihood of multiple segments of data, as defined by the vector of possible start-points (\code{tau}) and a single common end-point (\code{R}). This calculation assumes the shape is fixed as some specified value and sets the scale equal to its maximum likelihood estimate.
}
\examples{
shape = 1 # the shape parameter of the data
data = rgamma(n=100, shape=shape, scale=1/2)
sumx = pelt.gamma.sum(data)
tau = c(30,40,50) # start-points of possible segments
R = 70 # end-point of possible segments
pelt.gamma.cost(tau, R, sumx, shape=shape) # costs for each segment
shape = 1 # the shape parameter of the data
data = rgamma(n=100, shape=shape, scale=1/2)
sumx = pelt.gamma.sum(data)
tau = c(30,40,50) # start-points of possible segments
R = 70 # end-point of possible segments
pelt.gamma.cost(tau, R, sumx, shape=shape) # costs for each segment
}
