% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelt.R, R/pelt.bck.R
\name{pelt.gamma.cost.seglen}
\alias{pelt.gamma.cost.seglen}
\title{Title}
\usage{
pelt.gamma.cost.seglen(tau, R, sumx, shape = 1)

pelt.gamma.cost.seglen(tau, R, sumx, shape = 1)
}
\arguments{
\item{shape}{}

\item{shape}{}
}
\value{
A vector containing the cost of the different possible segments defined by \code{tau} and \code{R}.

A vector containing the cost of the different possible segments defined by \code{tau} and \code{R}.
}
\description{
Title

Title
}
