% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelt.R, R/pelt.bck.R
\name{pelt.gamma.sum}
\alias{pelt.gamma.sum}
\title{Calculate Summary Statistics used for Gamma likelihood calculation within PELT.}
\usage{
pelt.gamma.sum(data)

pelt.gamma.sum(data)
}
\arguments{
\item{data}{A vector of time-ordered observations to be analysed using PELT.}

\item{data}{A vector of time-ordered observations to be analysed using PELT.}
}
\value{
A vector containing the cumulative sums of the observations.

A vector containing the cumulative sums of the observations.
}
\description{
Calculates the necessary summary statistics which are to be used as part of the calculation of the Gamma likelihood for the PELT algorithm. Not typically used outside of this purpose.

Calculates the necessary summary statistics which are to be used as part of the calculation of the Gamma likelihood for the PELT algorithm. Not typically used outside of this purpose.
}
\details{
An internal function, not designed for use by the end-user.

An internal function, not designed for use by the end-user.
}
\examples{
data = rgamma(n=100, shape=1, scale=1/2)
pelt.gamma.sum(data)
data = rgamma(n=100, shape=1, scale=1/2)
pelt.gamma.sum(data)
}
