% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelt.R, R/pelt.bck.R
\name{pelt.norm.sum}
\alias{pelt.norm.sum}
\title{Calculate Summary Statistics used for Normal likelihood calculation within PELT.}
\usage{
pelt.norm.sum(data)

pelt.norm.sum(data)
}
\arguments{
\item{data}{A vector of time-ordered observations to be analysed using PELT.}

\item{data}{A vector of time-ordered observations to be analysed using PELT.}
}
\value{
A \code{3} x \code{length(data)} matrix. Row 1 contains the cumulative sums of the observations, row 2 contains the cumulative sums of the squared observations, and row 3 contains the cumulative sums of the squared de-meaned observations.

A \code{3} x \code{length(data)} matrix. Row 1 contains the cumulative sums of the observations, row 2 contains the cumulative sums of the squared observations, and row 3 contains the cumulative sums of the squared de-meaned observations.
}
\description{
Calculates the necessary summary statistics which are to be used as part of the calculation of the Normal likelihood for the PELT algorithm. Not typically used outside of this purpose.

Calculates the necessary summary statistics which are to be used as part of the calculation of the Normal likelihood for the PELT algorithm. Not typically used outside of this purpose.
}
\details{
An internal function, not designed for use by the end-user.

An internal function, not designed for use by the end-user.
}
\examples{
data = rnorm(100, 10, 3)
pelt.norm.sum(data)
data = rnorm(100, 10, 3)
pelt.norm.sum(data)
}
