% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pelt.R, R/pelt.bck.R
\name{pelt.norm.var.cost}
\alias{pelt.norm.var.cost}
\title{Calculate Normal likelihood of data segment, assuming fixed mean and variable variance.}
\usage{
pelt.norm.var.cost(tau, R, sumx)

pelt.norm.var.cost(tau, R, sumx)
}
\arguments{
\item{tau}{A vector of locations indicating the possible beginnings of a segment.}

\item{R}{A single location representing the end of the possible segment}

\item{sumx}{The summary statistics for the entire time series being analysed.}

\item{tau}{A vector of locations indicating the possible beginnings of a segment.}

\item{R}{A single location representing the end of the possible segment}

\item{sumx}{The summary statistics for the entire time series being analysed.}
}
\value{
A vector containing the cost of the different possible segments defined by \code{tau} and \code{R}.

A vector containing the cost of the different possible segments defined by \code{tau} and \code{R}.
}
\description{
Calculates the Normal likelihood of multiple segments of data, as defined by the vector of possible start-points (\code{tau}) and a single common end-point (\code{R}). This likelihood calculation assumes that the mean is fixed as \code{1}, and sets the variance equal to its maximum likelihood estimate.

Calculates the Normal likelihood of multiple segments of data, as defined by the vector of possible start-points (\code{tau}) and a single common end-point (\code{R}). This likelihood calculation assumes that the mean is fixed as \code{1}, and sets the variance equal to its maximum likelihood estimate.
}
\examples{
data = rnorm(100, 10, 3)
sumx = pelt.norm.sum(data)
tau = c(30,40,50) # start-points of possible segments
R = 70 # end-point of possible segments
pelt.norm.var.cost(tau,R,sumx) # costs for each segment
data = rnorm(100, 10, 3)
sumx = pelt.norm.sum(data)
tau = c(30,40,50) # start-points of possible segments
R = 70 # end-point of possible segments
pelt.norm.var.cost(tau,R,sumx) # costs for each segment
}
