% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cptmv.class.R
\docType{methods}
\name{plot,cptmv-method}
\alias{plot,cptmv-method}
\title{Title}
\usage{
\S4method{plot}{cptmv}(x, y, ...)
}
\arguments{
\item{x}{The x-axis values of the observations.}
}
\description{
Title
}
\examples{
a
}
