% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Quantile}
\alias{Quantile}
\title{Helper function to estimate the quantile of a sample distribution
Returns a vector of the x values less or equal to the selected quantile}
\description{
Helper function to estimate the quantile of a sample distribution
Returns a vector of the x values less or equal to the selected quantile
}
\details{
The function is taken from  https://github.com/RcppCore/Rcpp/issues/967
The authors are Rodney Sparapani and Dirk Eddelbuettel
}
