% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smop.for.asmop.R
\name{smop.for.asmop}
\alias{smop.for.asmop}
\title{Subset Multivariate Optimal Partitioning for Possible Changepoint Locations and Corresponding Subsets}
\usage{
smop.for.asmop(data, possible.cpts, possible.subsets, alpha, beta,
  min.dist, cost.func = norm.meanvar, initial.likelihood.value = 0,
  verbose = F)
}
\arguments{
\item{data}{An \code{n} x \code{p} matrix representing a length \code{n} time series containing observations of \code{p} variables.}

\item{possible.cpts}{A vector containing all possible changepoint locations, which will be used to generate the search-space of changepoint vectors.}

\item{possible.subsets}{A list of length \code{length(possible.cpts)}, where each element is a logical matrix containing \code{ncol(data)} columns. The rows of this matrix represent the different possible affected variable subsets for the corresponding element in \code{possible.cpts}.}

\item{alpha}{The variable-specific penalty, used to penalise the addition of a given changepoint into a given variable. A non-negative numeric value.}

\item{beta}{The multivariate penalty, used to penalise the addition of a changepoint into the model. A non-negative numeric value.}

\item{min.dist}{The minimum distance allowed between any two changepoints. Required to have an integer value of at least 2.}

\item{cost.func}{A function used to calculate the multivariate cost within SMOP. See details for more information.}

\item{initial.likelihood.value}{Numeric. The initial value of the likelihood.}

\item{verbose}{Logical. If \code{TRUE} then information regarding the changepoint vector check-list is printed during the algorithm.}
}
\value{
A list containing the following elements:
   \item{num.cpt.vecs}{The number of changepoint vectors within the search-space considered.}
   \item{cpt.vecs}{A matrix containing the optimal changepoint vectors for the series.}
   \item{like}{The value of the likelihood for the optimal set of changepoint vectors.}
   \item{cpts}{The optimal changepoint locations in the series.}
   \item{subsets}{A logical matrix containing the optimal affected variable subsets for each of the detected changepoints.}
}
\description{
This method performs the Subset Multivariate Optimal Partitioning (SMOP) algorithm for a given set of possible changepoint locations and given sets of possible corresponding affected variable subsets. The method is designed for use as part of the Approximate SMOP algorithm, with the sets of possible changepoint locations being obtained using PELT, and the affected variable subsets being obtained using hard or soft restriction.
}
\details{
The method takes a given set of possible changepoint locations and corresponding sets of possible affected variable subsets, and uses these to generate a search-space consisting of all possible changepoint vectors within the time series. An optimal partitioning search is used to obtain the optimal changepoint vector(s) from this search-space which minimise the penalised cost, as defined by \code{cost.func} and the penalties \code{alpha} and \code{beta}.

The possible changepoint locations in \code{checklist.cpts} are obtained by applying PELT to each variable (i.e. column) in \code{data}. The corresponding sets of possible affected variable subsets are obtained by applying either hard restriction or soft restriction (specified by the user) to the set of possible changepoint locations.

Values currently supported for the cost function \code{cost.func} include:
\tabular{ll}{
   \code{"norm.mean"} \tab Used for detecting changes in mean in Normally-distributed data. Assumes fixed variance parameters (\code{= 1}). The mean parameters are set to their maximum likelihood estimates. \cr
   \code{"norm.var"} \tab Used for detecting changes in variance in Normally-distributed data. Assumes fixed mean parameters (\code{= 0}). The variance parameters are set to their maximum likelihood estimates. \cr
   \code{"norm.meanvar"} \tab Used for detecting changes in both mean and variance in Normally-distributed data. The mean and variance parameters are set to their maximum likelihood estimates. \cr
   \code{"norm.mean.seglen"}, \code{"norm.var.seglen"}, \code{"norm.meanvar.seglen"} \tab Identical to \code{"norm.mean"}, \code{"norm.var"} and \code{"norm.meanvar"}, respectively, except these contain an additional log(segment length) penalty term in the likelihood for each variable. Designed for use when using the modified BIC penalty (Zhang and Siegmund, 2007) to penalise changes. \cr
}

This method is designed to be internal, and is not intended for explicit use by the end-user.
}
