% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_vars.R
\name{std_vars}
\alias{std_vars}
\title{std_vars: Standardize predictors}
\usage{
std_vars(pred, as_matrix = TRUE)
}
\arguments{
\item{pred}{An object of class SpatRaster or matrix or data.frame}

\item{as_matrix}{Logical. If TRUE a matrix will be returned}
}
\value{
A matrix or SpatRaster of standarized predictors
}
\description{
Standardizes predictors by converting them into z-scores.
}
\details{
if as_matrix is FALSE and  pred is a SpatRaster, the function would return
an SpatRaster object.
}
\examples{
pred_path <- system.file("extdata/G_layers", package = "smop") |>
                         list.files(full.names=TRUE)
pred <- terra::rast(pred_path)
std_bios <- smop::std_vars(pred,as_matrix=FALSE)
terra::plot(std_bios)
}
