% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmeasure.R
\name{vmeasure}
\alias{vmeasure}
\title{V-Measure}
\usage{
vmeasure(segmentation, true.cpts, true.subsets)
}
\arguments{
\item{segmentation}{A \code{cpt.mv} object produced using \code{smop} or \code{asmop}, representing an estimated set of changepoints and affected variable subsets.}

\item{true.cpts}{The vector of true changepoint locations.}

\item{true.subsets}{The matrix of true affected variable subsets corresponding to the true changepoints.}
}
\value{
The V-measure of the provided segmentation produced by SMOP/A-SMOP.
}
\description{
Calculates the V-measure of a subset-multivariate changepoint segmentation produced by \code{smop} or \code{asmop}.
}
