% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcMOP}
\alias{calcMOP}
\title{Compute mop serial version}
\arguments{
\item{x}{is the reference matrix.}

\item{y}{is the trasfer matrix.}
}
\value{
Returns a vector of mop values
}
\description{
Compute mop serial version
}
\details{
The unique difference with the function parcalcMOP
is that it prints the percentage of advance
}
